requirements.txt
numpy==1.11.2
PyStemmer==1.3.0
scikit-learn==0.18.1
beautifulsoup4==4.4.1
urllib3==1.13.1
